requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
depthai==2.15.1.0
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
idna==3.3
numpy==1.19.5
opencv-python==4.5.5.64
protobuf==3.19.4
psutil ==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
